+++ R/veg_layers-data.R
@@ -1,0 +1,14 @@
+#' @name veg_layers-data
+#' @aliases veg_layers
+#' @docType data
+#'
+#' @title Information on vegetation layers
+#'
+#' @description
+#' A data frame with a standard classification on vegetation layers.
+#'
+#' @format A data frame.
+#'
+#' @examples
+#' summary(veg_layers)
+"veg_layers"